Update for PSP match on `st0` `e_particles`
