+++ config/splat.pspeu.stst0.yaml
@@ -63,7 +63,7 @@ segments:
       - [0x12CB0, c, st0/st_update]
       - [0x131E8, c, st0_psp/e_particles_maria1]
       - [0x13770, c, st0_psp/e_particles_maria2]
-      - [0x13C00, c, st0_psp/e_particles]
+      - [0x13C00, c, st0/e_particles]
       - [0x14160, c, st0_psp/e_collect]
       - [0x15CC8, c, st0_psp/e_misc]
       - [0x16278, c, st0/collision]
@@ -81,7 +81,8 @@ segments:
       - [0x41BF0, .data, st0/e_lock_camera]
       - [0x41C08, .data, st0/e_red_door]
       - [0x41C20, .data, st0/st_update]
-      - [0x41C50, data]
+      - [0x41C50, .data, st0/e_particles]
+      - [0x41CD0, data]
       - [0x41F98, .data, st0/collision]
       - [0x42308, data, e_room_fg]
       - [0x423A8, data]
