+++ config/symbols.pspeu.stst0.txt
@@ -67,6 +67,8 @@ ST0_EntityRedDoor = 0x09249110;
 Random = 0x0924A330;
 Update = 0x0924A370;
 UpdateStageEntities = 0x0924A750;
+func_psp_0923AD68 = 0x0924A868;
+func_psp_0923B2F0 = 0x0924ADF0;
 EntitySoulStealOrb = 0x0924B280;
 PrizeDropFall = 0x0924B7E0;
 PrizeDropFall2 = 0x0924B8B8;
@@ -131,6 +133,9 @@ g_EInitSecretStairs = 0x09252D60;
 ST0_RedDoorTiles = 0x09252D70;
 g_Rooms = 0x09257350;
 ST0_BackgroundBlockInit = 0x092791F0;
+g_ESoulStealOrbAngles = 0x092792D0;
+g_ESoulStealOrbSprt = 0x092792E0;
+g_ESoulStealOrbAnim = 0x09279340;
 aluric_subweapons_idx = 0x09279420;
 aluric_subweapons_id = 0x09279438;
 c_GoldPrizes = 0x09279450;
