--- src/st/st0_psp/e_particles.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "common.h"
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_particles", EntitySoulStealOrb);
