+++ src/st/st0_psp/e_particles_maria1.c
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-#include "common.h"
+#include "../st0/st0.h"
 
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_particles_maria1", func_pspeu_0924A868);
+#include "../e_particles_maria1.h"
